+++ v4/analyzeGaussianAgent.m
@@ -1,7 +1,7 @@
 function analyzeGaussianAgent(
   X,                  # the indicators dump
-  EPSMU = 0.5,        # the averege range of delta mu
-  EPSHS = 0.2,        # the average range of delta h sigma
+  EPSMU = [10e-3 62.5e-3 250e-3],        # the averege range of delta mu
+  EPSHS = [347e-3 347e-3 347e-3],        # the average range of delta h sigma
   K0 = 0.7,           # the K threshold for C1 class
   EPS = 10e-6,         # the minimu J value to be considered
   PRC = [50 : 10 : 90]', # the percentiles in the chart
@@ -12,39 +12,23 @@ function analyzeGaussianAgent(
   # Number of steps, number of values
   [n m] = size(X);
   # Number of actors
-  NA = floor((m - 4 ) / NCOLA);
+  NA = floor((m - 6) / NCOLA);
   DR = 1:3;
 
-  # Compute J value of network (sum over the J values of critics and actors)
+  # Extract the state value estimation for the state
   V0 = X(:, 1);
+  # Extract the target state value estimation for the state
   VSTAR = X(:, 2);
+  # Extract the state value estimation for the next state
   V1 = X(:, 3);
+  # Extract the average state value
   RPI = X(:, 4);
   TD = (VSTAR - V0) .^ 2;
-  J = TD;
-  J1 = (VSTAR - V1) .^ 2;
-
-  for actor = 0 : NA - 1
-    j = actor * NCOLA + 5;
-    HMU = X(:, j + 1);
-    HMUSTAR = X(:, j + 2);
-    HMU1 = X(:, j + 3);
-    
-    HS = X(:, j + 5);
-    HSSTAR = X(:, j + 6);
-    HS1 = X(:, j + 7);
-
-    JHMU = (HMUSTAR - HMU) .^ 2;
-    JHMU1 = (HMUSTAR - HMU1) .^ 2;
-
-    JHS = (HSSTAR - HS) .^ 2;
-    JHS1 = (HSSTAR - HS1) .^ 2;
-
-    J = J + JHMU + JHS;
-    J1 = J1 + JHMU1 + JHS1;
-   
-  endfor
-
+  # Compute J value of network (mean square error over outputs before training)
+  J = X(:, 5);
+  # Compute J' value of network  (mean square error over outputs after training)
+  J1 = X(:, 6);
+ 
   # Filter the J values greater than threshold EPS
   VI = find(abs(J) >= EPS);
   J = J(VI, 1);
@@ -53,39 +37,40 @@ function analyzeGaussianAgent(
   K = J1 ./ J;
   [TDTREND, TDP, TDMODE] = regression(TD);
   [RTREND, RP, RMODE] = regression(RPI);
-  
+
   # Number of invalid steps
   C0 = sum(K > 1);
-  # Number of otpimizing steps
+  # Number of optimizing steps
   C1 = sum(K <= 1 & K > K0);
   # Number of optimizied steps
   C2 = n - C0 - C1;
-  
+
   muChart = {};
   sigmaChart = {};
   for actor = 0 : NA - 1
-    j = actor * NCOLA + 5;
-    alphaMu = X(1, j);
-    # Computes the J values
-    HMU = X(:, j + 1);
-    HMUSTAR = X(:, j + 2);
+    j = actor * NCOLA;
+    alphaMu = X(1, j + 7);
     
-    DHMU = HMUSTAR - HMU;
-    DHMU2 = DHMU .^ 2;
+    # Computes mu
+    MU = X(:, j + 8);
+    MUSTAR = X(:, j + 9);
+    
+    DMU = MUSTAR - MU;
+    DMU2 = DMU .^ 2;
 
-    muChart{actor + 1 , 1} = DHMU2;
+    muChart{actor + 1 , 1} = DMU2;
 
     # Compute alpha for percetile
-    PCHMU = prctile(sqrt(DHMU2), PRC);
-    muChart{actor + 1, 2} = EPSMU ./ PCHMU * alphaMu;
+    PCMU = prctile(sqrt(DMU2), PRC);
+    muChart{actor + 1, 2} = EPSMU(actor + 1) ./ PCMU * alphaMu;
 
     # Compute mean
-    muChart{actor + 1, 3} = ones(size(PRC, 1), 1) * EPSMU ./ sqrt(mean(DHMU2)) * alphaMu;
+    muChart{actor + 1, 3} = ones(size(PRC, 1), 1) * EPSMU(actor + 1) ./ sqrt(mean(DMU2)) * alphaMu;
 
-    alphaHs = X(1, j + 4);
+    alphaHs = X(1, j + 11);
     # Computes the J values
-    HS = X(:, j + 5);
-    HSSTAR = X(:, j + 6);
+    HS = X(:, j + 12);
+    HSSTAR = X(:, j + 13);
     
     DHS = HSSTAR - HS;
     DHS2 = DHS .^ 2;
@@ -94,10 +79,10 @@ function analyzeGaussianAgent(
 
     # Compute alpha for percetile
     PCHS = prctile(sqrt(DHS2), PRC);
-    sigmaChart{actor + 1, 2} = EPSHS ./ PCHS * alphaHs;
+    sigmaChart{actor + 1, 2} = EPSHS(actor + 1) ./ PCHS * alphaHs;
 
     # Compute mean
-    sigmaChart{actor + 1, 3} = ones(size(PRC, 1), 1) * EPSHS ./ sqrt(mean(DHS2)) * alphaHs;
+    sigmaChart{actor + 1, 3} = ones(size(PRC, 1), 1) * EPSHS(actor + 1) ./ sqrt(mean(DHS2)) * alphaHs;
   endfor
   
   NR = 4;
@@ -149,14 +134,14 @@ function analyzeGaussianAgent(
     ylabel(sprintf("\\alpha_\\mu %d", actor));
 
     subplot(NR, NC, col + 2 * NC);
-    hist(muChart{actor + 1, 1}, BINS);
+    hist(sigmaChart{actor + 1, 1}, BINS);
     grid on;
     title(sprintf("J h%d distribution", actor));
     xlabel(sprintf("J h %d", actor));
     ylabel("# samples");
 
     subplot(NR, NC, col + 3 * NC);
-    autoplot(PRC, [muChart{actor + 1, 2}, muChart{actor + 1, 3}]);
+    autoplot(PRC, [sigmaChart{actor + 1, 2}, sigmaChart{actor + 1, 3}]);
     grid on;
     grid minor on;
     title(sprintf("\\alpha_h %d", actor));